+++ OpenMarket/OpenMarket/Extensions/Double+Extensions.swift
@@ -1,5 +1,5 @@
 //
-//  Int+Extensions.swift
+//  Double+Extensions.swift
 //  OpenMarket
 //
 //  Created by 데릭, 케이, 수꿍. 
@@ -7,7 +7,7 @@
 
 import Foundation
 
-extension Int {
+extension Double {
     func numberFormatter() -> String {
         let numberFormatter = NumberFormatter()
         numberFormatter.numberStyle = .decimal
